fix(platform): split skeleton components to enable Fast Refresh
Fast Refresh was performing full reloads because component files
exported multiple React components (main + skeleton). Split each
skeleton into its own file.

Files split:
- HomePageStats → HomePageStatsSkeleton
- StudentAthleteGrid → StudentAthleteGridSkeleton
- ReviewsList → ReviewsListSkeleton
- AddReviewFormWrapper → AddReviewFormSkeleton
- MyJobsTabs → MyJobsTabsSkeleton
- DashboardJobTabs → DashboardJobTabsSkeleton
- AdminUserTabs → AdminUserTabsSkeleton
