+++ package.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/mikelong10"
   },
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --webpack",
     "build": "next build --webpack",
     "start": "next start",
     "lint": "biome check --write .",
