+++ src/app/(app)/(personal)/jobs/MyJobsTabsSkeleton.tsx
@@ -1,0 +1,33 @@
+export function MyJobsTabsSkeleton() {
+  return (
+    <div className="w-full">
+      <div className="bg-muted mb-4 flex w-full gap-1 rounded-lg p-1">
+        <div className="bg-background h-9 flex-1 animate-pulse rounded-md" />
+        <div className="bg-muted h-9 flex-1 rounded-md" />
+        <div className="bg-muted h-9 flex-1 rounded-md" />
+        <div className="bg-muted h-9 flex-1 rounded-md" />
+      </div>
+      <div className="flex flex-col gap-4">
+        <div className="bg-muted h-5 w-64 animate-pulse rounded" />
+        <div className="grid grid-cols-1 gap-4 md:grid-cols-2 xl:grid-cols-3">
+          {[1, 2, 3].map((i) => (
+            <div
+              key={i}
+              className="bg-card flex flex-col gap-4 rounded-lg border p-4"
+            >
+              <div className="flex flex-col gap-2">
+                <div className="bg-muted h-6 w-40 animate-pulse rounded" />
+                <div className="bg-muted h-4 w-24 animate-pulse rounded" />
+              </div>
+              <div className="flex flex-col gap-2">
+                <div className="bg-muted h-4 w-full animate-pulse rounded" />
+                <div className="bg-muted h-4 w-2/3 animate-pulse rounded" />
+              </div>
+              <div className="bg-muted h-6 w-20 animate-pulse rounded-full" />
+            </div>
+          ))}
+        </div>
+      </div>
+    </div>
+  );
+}
